pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "graphicc"
version = "0.1.0"
[tool.poetry.dependencies]
python = "^3.7"
termcolor = "^1.1"
ansicolors = "^1.1"
[tool.poetry.dev-dependencies]
